Replace rendering with Water templates

+++ src/main/java/co/q64/vpn/net/SparkServer.java
@@ -2,6 +2,7 @@ package co.q64.vpn.net;
 
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
@@ -18,13 +19,12 @@ import java.util.concurrent.TimeUnit;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
-import org.eclipse.egit.github.core.client.GitHubClient;
-import org.eclipse.egit.github.core.service.UserService;
 import org.json.JSONObject;
 
 import spark.Service;
 import spark.Session;
 import spark.Spark;
+import spark.template.water.WaterTemplateEngine;
 import co.q64.vpn.api.config.Config;
 import co.q64.vpn.api.database.Database;
 import co.q64.vpn.api.log.Logger;
@@ -32,43 +32,58 @@ import co.q64.vpn.api.net.Server;
 import co.q64.vpn.objects.CodeData;
 import co.q64.vpn.objects.CodeData.CodeUsage;
 import co.q64.vpn.objects.UserData;
-import co.q64.vpn.page.AccountPageRenderer;
-import co.q64.vpn.page.BasicHTMLComponents;
-import co.q64.vpn.page.InvalidPageRenderer;
-import co.q64.vpn.page.InvitePageRenderer;
-import co.q64.vpn.page.LoginPageRenderer;
+import co.q64.vpn.page.AccountPage;
+import co.q64.vpn.page.DemoPage;
+import co.q64.vpn.page.ErrorPage;
+import co.q64.vpn.page.ForceTosPage;
+import co.q64.vpn.page.InfoPage;
+import co.q64.vpn.page.InvitePage;
+import co.q64.vpn.page.LoginPage;
+import co.q64.vpn.page.LogoutPage;
+import co.q64.vpn.page.TosPage;
+import co.q64.vpn.page.VerifyPage;
 import co.q64.vpn.util.IPSECUpdater;
 import co.q64.vpn.util.PayPalAPI;
 import co.q64.vpn.util.TimeUtil;
 
-import com.github.scribejava.apis.GitHubApi;
-import com.github.scribejava.apis.GoogleApi20;
-import com.github.scribejava.core.builder.ServiceBuilder;
-import com.github.scribejava.core.model.OAuth2AccessToken;
-import com.github.scribejava.core.model.Token;
-import com.github.scribejava.core.oauth.OAuth20Service;
-import com.google.gson.JsonObject;
+import com.google.firebase.FirebaseApp;
+import com.google.firebase.FirebaseOptions;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseToken;
+import com.google.firebase.tasks.OnSuccessListener;
 
 @Singleton
 public class SparkServer implements Server {
 	private @Inject Config config;
-	private @Inject LoginPageRenderer loginPage;
-	private @Inject InvitePageRenderer invitePage;
-	private @Inject AccountPageRenderer accountPage;
-	private @Inject InvalidPageRenderer invalidPage;
 	private @Inject Database database;
 	private @Inject TimeUtil time;
 	private @Inject IPSECUpdater ipsec;
 	private @Inject Logger logger;
 	private @Inject PayPalAPI paypal;
 
-	private OAuth20Service githubService;
+	private FirebaseAuth auth;
 	private Map<String, Long> lastRequest = new HashMap<String, Long>();
 	private ScheduledThreadPoolExecutor pool = new ScheduledThreadPoolExecutor(1);
 
 	@Inject
 	public void init() {
-		githubService = new ServiceBuilder().apiKey(config.getOAuthID()).apiSecret(config.getOAuthSecret()).callback(config.getServerURL() + "/callback/gh").scope("user:email,repo").build(GitHubApi.instance());
+		File firebase = new File(Paths.get("firebase.json").toUri());
+		if (!firebase.exists()) {
+			logger.error("Could not find firebase config at " + firebase.getAbsolutePath());
+			return;
+		}
+		//formatter:off
+		FirebaseOptions options = null;
+		try {
+			options = new FirebaseOptions.Builder()
+			  .setServiceAccount(new FileInputStream(firebase))
+			  .build();
+		} catch (FileNotFoundException e1) {
+			e1.printStackTrace();
+		}
+		//formatter:on
+		FirebaseApp.initializeApp(options);
+		auth = FirebaseAuth.getInstance();
 		pool.scheduleAtFixedRate(() -> {
 			Set<String> remove = new HashSet<String>();
 			for (Entry<String, Long> e : lastRequest.entrySet()) {
@@ -96,19 +111,94 @@ public class SparkServer implements Server {
 
 		Spark.secure(config.getJKSPath(), config.getJKSPassword(), null, null);
 		Spark.port(443);
+
+		Spark.get("/login", (request, response) -> {
+			return WaterTemplateEngine.render(new LoginPage(), request);
+		}, WaterTemplateEngine.waterEngine());
+
+		Spark.get("/verify", (request, response) -> {
+			return WaterTemplateEngine.render(new VerifyPage(), request);
+		}, WaterTemplateEngine.waterEngine());
+
+		Spark.post("/token/*", (request, response) -> {
+			Session session = request.session();
+			String token = request.queryParams("user");
+			//logger.info("Auth start " + token);
+				auth.verifyIdToken(token).addOnSuccessListener(new OnSuccessListener<FirebaseToken>() {
+
+					@Override
+					public void onSuccess(FirebaseToken token) {
+						session.attribute("token", token);
+						logger.info("Auth accepted for " + token.getUid());
+					}
+				});
+				return "";
+			});
+
+		Spark.get("/access", (request, response) -> {
+			Session s = request.session();
+			FirebaseToken token = s.attribute("token");
+			UserData data = s.attribute("data");
+			if (token == null || data == null) {
+				response.redirect("/");
+				return null;
+			}
+			return WaterTemplateEngine.render(new InvitePage(), request);
+		}, WaterTemplateEngine.waterEngine());
+
+		Spark.get("/account", (request, response) -> {
+			Session s = request.session();
+			FirebaseToken token = s.attribute("token");
+			UserData data = s.attribute("data");
+			if (token == null || data == null || Boolean.valueOf(data.getIsNew()) || !Boolean.valueOf(data.getIsTos())) {
+				response.redirect("/");
+				return null;
+			}
+			return WaterTemplateEngine.render(new AccountPage(data, token.getName(), time.getAccountTerminationTime(data)), request);
+		}, WaterTemplateEngine.waterEngine());
+
+		Spark.get("/forcetos", (request, response) -> {
+			Session s = request.session();
+			FirebaseToken token = s.attribute("token");
+			UserData data = s.attribute("data");
+			if (token == null || data == null || Boolean.valueOf(data.getIsTos())) {
+				response.redirect("/");
+				return null;
+			}
+			return WaterTemplateEngine.render(new ForceTosPage(), request);
+		}, WaterTemplateEngine.waterEngine());
+
+		Spark.get("/conftos", (request, response) -> {
+			Session s = request.session();
+			FirebaseToken token = s.attribute("token");
+			UserData data = s.attribute("data");
+			if (token == null || data == null || Boolean.valueOf(data.getIsTos())) {
+				response.redirect("/");
+				return null;
+			}
+			data.setIsTos(String.valueOf(true));
+			response.redirect("/");
+			return null;
+		});
+
+		Spark.get("robots.txt", (request, response) -> {
+			return "User-agent: *\nDisallow: /";
+		});
+
+		Spark.get("/tos", (request, response) -> {
+			return WaterTemplateEngine.render(new TosPage(), request);
+		}, WaterTemplateEngine.waterEngine());
+
 		Spark.get("/", (request, response) -> {
 			Session s = request.session();
-			OAuth2AccessToken token = s.attribute("token");
+			FirebaseToken token = s.attribute("token");
 			UserData data = s.attribute("data");
 			if (token == null) {
-				return loginPage.render(githubService.getAuthorizationUrl());
+				response.redirect("/verify");
+				return null;
 			}
-
 			if (data == null) {
-				GitHubClient client = new GitHubClient();
-				client.setOAuth2Token(token.getAccessToken());
-				UserService userService = new UserService(client);
-				String id = userService.getUser().getLogin();
+				String id = token.getUid();
 				database.disconnect(id);
 				database.queryData(id);
 				UserData userData = database.getData(UserData.class, id);
@@ -121,9 +211,23 @@ public class SparkServer implements Server {
 				return null;
 			}
 			if (Boolean.valueOf(data.getIsNew())) {
-				return invitePage.render(data);
+				if (s.attribute("preauth") != null && Boolean.valueOf(s.attribute("preauth"))) {
+					data.setIsNew(String.valueOf(false));
+				} else {
+					response.redirect("/access");
+					return null;
+				}
 			}
-			return accountPage.render(data);
+			if (!Boolean.valueOf(data.getIsTos())) {
+				response.redirect("/forcetos");
+				return null;
+			}
+			response.redirect("/account");
+			return null;
+		});
+
+		Spark.exception(Exception.class, (exception, request, response) -> {
+			exception.printStackTrace();
 		});
 
 		Spark.post("/invite", (request, response) -> {
@@ -143,7 +247,9 @@ public class SparkServer implements Server {
 			CodeData data = database.getData(CodeData.class, code);
 			if (Boolean.valueOf(data.getIsValid()) && data.getUsage().equals(CodeUsage.ACCESS.name())) {
 				ud.setIsNew(String.valueOf(false));
-				database.deleteData(data);
+				if (!Boolean.valueOf(data.getIsPerm())) {
+					database.deleteData(data);
+				}
 				database.disconnect(code);
 				response.redirect("/");
 				return null;
@@ -152,6 +258,27 @@ public class SparkServer implements Server {
 			response.redirect("/invalid");
 			return null;
 		});
+
+		Spark.get("/preauth/*", (request, response) -> {
+			Session s = request.session();
+			String token = request.queryParams("token");
+			if (token == null || token.isEmpty()) {
+				return WaterTemplateEngine.render(new ErrorPage("That token is invalid"), request);
+			}
+			database.disconnect(token);
+			database.queryData(token);
+			CodeData data = database.getData(CodeData.class, token);
+			if (Boolean.valueOf(data.getIsValid()) && data.getUsage().equals(CodeUsage.ACCESS.name())) {
+				if (!Boolean.valueOf(data.getIsPerm())) {
+					database.deleteData(data);
+				}
+				database.disconnect(token);
+				s.attribute("preauth", String.valueOf(true));
+				return WaterTemplateEngine.render(new DemoPage(), request);
+			}
+			database.disconnect(token);
+			return WaterTemplateEngine.render(new ErrorPage("That token is invalid"), request);
+		}, WaterTemplateEngine.waterEngine());
 
 		Spark.post("/redeem", (request, response) -> {
 			Session s = request.session();
@@ -175,13 +302,21 @@ public class SparkServer implements Server {
 				}
 				time += TimeUnit.DAYS.toMillis(data.getAmount());
 				ud.setEndTime(time);
-				if (ipsec.get(ud.getId()) == null) {
+				if (ipsec.get(ud.getUsername()) == null) {
+					if (ud.getUsername().equals("default")) {
+						StringBuilder user = new StringBuilder();
+						for (int i = 0; i < 8; i++) {
+							char c = (char) (ThreadLocalRandom.current().nextInt(26) + 'a');
+							user.append(c);
+						}
+						ud.setUsername(user.toString());
+					}
 					StringBuilder pass = new StringBuilder();
 					for (int i = 0; i < 8; i++) {
 						char c = (char) (ThreadLocalRandom.current().nextInt(26) + 'a');
 						pass.append(c);
 					}
-					ipsec.update(ud.getId(), pass.toString());
+					ipsec.update(ud.getUsername(), pass.toString());
 					ud.setPass(pass.toString());
 				}
 				ipsec.updateNow();
@@ -300,12 +435,12 @@ public class SparkServer implements Server {
 		});
 
 		Spark.get("/invalid", (request, response) -> {
-			return invalidPage.render("That was an invalid code!");
-		});
+			return WaterTemplateEngine.render(new ErrorPage("That was an invalid code!"), request);
+		}, WaterTemplateEngine.waterEngine());
 
-		Spark.get("/session", (request, response) -> {
-			return invalidPage.render("Your session was ended due to inactivity");
-		});
+		Spark.get("/invalid", (request, response) -> {
+			return WaterTemplateEngine.render(new ErrorPage("Your session was ended due to inactivity"), request);
+		}, WaterTemplateEngine.waterEngine());
 
 		Spark.get("/logout", (request, response) -> {
 			Session s = request.session();
@@ -315,17 +450,17 @@ public class SparkServer implements Server {
 			}
 			s.removeAttribute("token");
 			s.removeAttribute("data");
-			response.redirect("/");
-			return null;
-		});
+			return WaterTemplateEngine.render(new LogoutPage(), request);
+		}, WaterTemplateEngine.waterEngine());
 
 		Spark.get("/terminated", (request, response) -> {
 			Session s = request.session();
 			s.removeAttribute("token");
 			s.removeAttribute("data");
-			return BasicHTMLComponents.BEGIN + "Account terminated" + BasicHTMLComponents.END;
-		});
+			return WaterTemplateEngine.render(new ErrorPage("Your account has been terminated"), request);
+		}, WaterTemplateEngine.waterEngine());
 
+		/*
 		Spark.get("/callback/*", (request, response) -> {
 			Session s = request.session();
 			Object token = s.attribute("token");
@@ -340,16 +475,17 @@ public class SparkServer implements Server {
 			response.redirect("/");
 			return null;
 		});
+		*/
 
 		Spark.get("/code/*", (request, response) -> {
 			String name = request.queryParams("name");
 			String time = request.queryParams("time");
 			if (time == null) {
-				return BasicHTMLComponents.BEGIN + "Your new access code is: " + name + BasicHTMLComponents.END;
+				return WaterTemplateEngine.render(new InfoPage("New access code", name), request);
 			} else {
-				return BasicHTMLComponents.BEGIN + "Your new time code for " + time + " days is: " + name + BasicHTMLComponents.END;
+				return WaterTemplateEngine.render(new InfoPage("New time code for " + time + " days", name), request);
 			}
-		});
+		}, WaterTemplateEngine.waterEngine());
 
 		// Paypal
 
@@ -379,6 +515,24 @@ public class SparkServer implements Server {
 				}
 				newtime += TimeUnit.DAYS.toMillis(time * 7);
 				ud.setEndTime(newtime);
+				if (ipsec.get(ud.getUsername()) == null) {
+					if (ud.getUsername().equals("default")) {
+						StringBuilder user = new StringBuilder();
+						for (int i = 0; i < 8; i++) {
+							char c = (char) (ThreadLocalRandom.current().nextInt(26) + 'a');
+							user.append(c);
+						}
+						ud.setUsername(user.toString());
+					}
+					StringBuilder pass = new StringBuilder();
+					for (int i = 0; i < 8; i++) {
+						char c = (char) (ThreadLocalRandom.current().nextInt(26) + 'a');
+						pass.append(c);
+					}
+					ipsec.update(ud.getUsername(), pass.toString());
+					ud.setPass(pass.toString());
+				}
+				ipsec.updateNow();
 				logger.info("Added time");
 				return new JSONObject().toString();
 			}
@@ -387,11 +541,11 @@ public class SparkServer implements Server {
 		});
 
 		Spark.get("/paypaldone", (request, response) -> {
-			return invalidPage.render("You have added time to your account.");
-		});
+			return WaterTemplateEngine.render(new InfoPage("Payment Successful", "Time has been added to your account"), request);
+		}, WaterTemplateEngine.waterEngine());
 
 		Spark.get("/paypalfail", (request, response) -> {
-			return invalidPage.render("Something went wrong during the transaction and you account has not been charged.");
-		});
+			return WaterTemplateEngine.render(new InfoPage("Payment Failure", "Something went wrong during the transaction and you account has not been charged"), request);
+		}, WaterTemplateEngine.waterEngine());
 	}
 }
